Enough for tonight. Sort of works
The abandon match is passing the opponent_id instead of the username, and its all a bit messy right now, but I am going to stop for tonight.
I should go through and refactor soon as well as get the abandoning match to display something on screen, leave the lobby, and navigate to landing page.